% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_chr}
\alias{gene_chr}
\title{Gene chromosome location}
\format{
A data frame with 60571 rows and 3 variables:
\describe{
\item{chromosome_name}{chromosome}
\item{ensembl_gene_id}{Ensembl gene ID of gene}
\item{hgnc_symbol}{HGNC symbol of gene}
}
}
\usage{
gene_chr
}
\description{
A dataset containing two formats of gene names and associated
chromosome number.
}
\keyword{datasets}
